eeML

Documentation for eeML.

eeML.LinearRegression

eeML.LinearRegression

eeML.train!

eeML.LinearRegression — Type

LinearRegression

A container for a linear regression model. Contains the predictor matrix, the targets, the coefficients, and the standard errors of the coefficients.

Fields

𝐗::Matrix{<:Real}: the predictor matrix

y::AbstractVector{<:Real}: the target vector

β̂::Vector{<:Real}: the estimated coefficients

SEᵦ::Vector{<:Real}: the standard errors of the estimated coefficients

eeML.LinearRegression — Method

LinearRegression

A constructor function for a new LinearRegression object

Arguments

𝐗: the predictor matrix

y: the target vector

eeML.train! — Method

train!

In-place training of a linear regression model

Arguments

model::LinearRegression: a LinearRegression model (which includes a predictor matrix and a target vector) to be fit